Rename file to match the main type
